Refactor loaders types and interfaces
