+++ packages/loaders/.size-snapshot.json
@@ -1,13 +1,13 @@
 {
   "index.cjs.js": {
-    "bundled": 25204,
-    "minified": 18944,
-    "gzipped": 5011
+    "bundled": 25222,
+    "minified": 18954,
+    "gzipped": 5022
   },
   "index.esm.js": {
-    "bundled": 23420,
-    "minified": 17364,
-    "gzipped": 4867,
+    "bundled": 23438,
+    "minified": 17374,
+    "gzipped": 4879,
     "treeshaked": {
       "rollup": {
         "code": 14239,
