+++ packages/loaders/src/elements/Dots.tsx
@@ -5,11 +5,12 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { useEffect, useRef, useContext, SVGAttributes, forwardRef } from 'react';
+import React, { useEffect, useRef, useContext, forwardRef } from 'react';
 import PropTypes from 'prop-types';
 import { ThemeContext } from 'styled-components';
 import { useSchedule } from '@zendeskgarden/container-schedule';
 import { useDocument } from '@zendeskgarden/react-theming';
+import { IDotsProps } from '../types';
 import {
   StyledDotsCircleOne,
   StyledDotsCircleTwo,
@@ -19,17 +20,6 @@ import {
 } from '../styled';
 
 const COMPONENT_ID = 'loaders.dots';
-
-export interface IDotsProps extends SVGAttributes<SVGSVGElement> {
-  /** Sets the height and width in pixels. Inherits the parent's font size by default. */
-  size?: string | number;
-  /** Sets the fill color. Inherits the parent's `color` by default. */
-  color?: string;
-  /** Sets the length of the animation cycle in milliseconds **/
-  duration?: number;
-  /** Delays displaying the loader to prevent a render flash during normal loading times **/
-  delayMS?: number;
-}
 
 /**
  * @extends SVGAttributes<SVGSVGElement>
