+++ packages/loaders/src/elements/Skeleton.tsx
@@ -5,18 +5,10 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { forwardRef, HTMLAttributes } from 'react';
+import React, { forwardRef } from 'react';
 import PropTypes from 'prop-types';
+import { ISkeletonProps } from '../types';
 import { StyledSkeleton } from '../styled';
-
-export interface ISkeletonProps extends HTMLAttributes<HTMLDivElement> {
-  /** Sets the width as a percentage of the the parent element's width */
-  width?: string;
-  /** Sets the height as a percentage of parent element's height if the height is not already inherited by `line-height` */
-  height?: string;
-  /** Inverts the color for use on dark backgrounds */
-  isLight?: boolean;
-}
 
 /**
  * @extends HTMLAttributes<HTMLDivElement>
